gestion erreur presque finie - asm fini mais pas norme/leaks - arene pas commencee

+++ tester/test.sh
@@ -3,12 +3,20 @@
 program=../asm
 map_good=0
 map_total=0
+go=0
 
 echo "\033[0;33mBienvenue dans le checker - auteur : gmoindro & rbeaufre\033[0m\n"
 
-mkdir 42 our
+if [ -d "42" ];then
+	echo "Next time, let the program run until the end or put some champions in it\n"
+else
+	mkdir 42 our
+fi
+
 for rep in $*
 do
+	go=$((1))
+	echo
 	l=$((${#rep} - 2))
 	map_valid=${rep:0:$l}
 	map_total=$(($map_total + 1))
@@ -28,6 +36,7 @@ do
 				echo "\033[0;32m\t[OK] $map_valid.s \033[0m"
 				map_good=$(($map_good + 1))
 			else
+				cmp 42/$x.cor our/$x.cor
 				echo "\033[0;31m\t[KO] $map_valid.s\t\t#problem inside the .cor\033[0m"
 			fi
 		fi
@@ -42,6 +51,10 @@ do
 		fi
 	fi
 done
-rm -r 42 our
 
-echo "\n\033[0;35mResult : [$map_good / $map_total] \033[0m"
+if [ $go != 0 ]; then
+	rm -r 42 our
+	echo "\n\033[0;35mResult : [$map_good / $map_total] \033[0m"
+else
+	echo "Write in test.sh the path to your asm\nusage: sh test.sh [file ...]"
+fi
